Refactor to make use of ndarray
This is the standard crate for dealing with multi-dimensional arrays.

In the long term, I'm not sure we really want ndarray as a dependency.
In particular, it's a fairly heavy dependency and I'm not sure it will
let us get the most from SIMD. I also suspect it may introduce some
problems if we wanted to try to use GPU support (this latter point
is **MUCH** less important, and as I understand it, most issues go away
if we don't compile the features that use the standard library).

In order to keep moving foward, I think the compromise is to use
ndarray for now, and plan to create an ArrayView type with the limited
required features in the future (if we think it is going to be
important).