Refactor documentation
The resulting documentation is shorter in several places, but I've tried to only delete things that were either redundant or seemed like one-off messages for my own comprehension.  This was a useful excercise to make sure I understood it, but I think the result will also be am improvement for other developers.