Introduce `OnePointBinned`
This primarily exists to serve as a concrete example of how the machinery
for driving parallel reductions can be used to perform a (conceptually)
simple calculation. This is useful in 2 contexts:
1. It's a pedagocial resource to help illustrating how the machinery
   actually works.
2. Because calculating a binned mean is conceptually simple, its really
   easy to implement a simple version of that logic without the
   parallelism machinery. This makes the type very useful for actually
   testing parallelism machinery

(The docstring also highlights a practical use-case)

The goal is to use this as a simpler, less contrived replacement for the
contents of `pairstat_nostd_internal/src/reduce_sample/*`. To that end,
we probably also want to implement another type that performs an
unbinned reduction, so we can have an example using the
`team.calccontribs_combine_apply` method (in practice, we would
implement the logic as though there is always 1 bin)